Making panel works on test files.
