+++ scripts/new_friends_nodes_edges.js
@@ -1,0 +1,23 @@
+var links = [
+  {source: nodes[Vivian], target: nodes[Alexandra], friendship: 'friends', edge_betweenness_centrality: '0.0378787878788'},
+  {source: nodes[Vivian], target: nodes[Eric], friendship: 'friends', edge_betweenness_centrality: '0.0934343434343'},
+  {source: nodes[Vivian], target: nodes[Alice], friendship: 'friends', edge_betweenness_centrality: '0.0959595959596'},
+  {source: nodes[Rob], target: nodes[Bob], friendship: 'friends', edge_betweenness_centrality: '0.0744949494949'},
+  {source: nodes[Rob], target: nodes[Asma], friendship: 'friends', edge_betweenness_centrality: '0.166666666667'},
+  {source: nodes[Rob], target: nodes[David], friendship: 'enemies', edge_betweenness_centrality: '0.0989898989899'},
+  {source: nodes[Rob], target: nodes[Alice], friendship: 'friends', edge_betweenness_centrality: '0.155303030303'},
+  {source: nodes[Rob], target: nodes[Jack], friendship: 'friends', edge_betweenness_centrality: '0.129292929293'},
+  {source: nodes[Alice], target: nodes[Alexandra], friendship: 'friends', edge_betweenness_centrality: '0.128787878788'},
+  {source: nodes[Alice], target: nodes[Jack], friendship: 'friends', edge_betweenness_centrality: '0.160353535354'},
+  {source: nodes[George], target: nodes[Bob], friendship: 'friends', edge_betweenness_centrality: '0.0522727272727'},
+  {source: nodes[George], target: nodes[Jack], friendship: 'friends', edge_betweenness_centrality: '0.0681818181818'},
+  {source: nodes[George], target: nodes[Carlos], friendship: 'friends', edge_betweenness_centrality: '0.0383838383838'},
+  {source: nodes[George], target: nodes[Eric], friendship: 'friends', edge_betweenness_centrality: '0.0679292929293'},
+  {source: nodes[Jack], target: nodes[Eric], friendship: 'friends', edge_betweenness_centrality: '0.116414141414'},
+  {source: nodes[Jack], target: nodes[Isaac], friendship: 'enemies', edge_betweenness_centrality: '0.166666666667'},
+  {source: nodes[Jack], target: nodes[Bob], friendship: 'friends', edge_betweenness_centrality: '0.0737373737374'},
+  {source: nodes[Jack], target: nodes[Carlos], friendship: 'enemies', edge_betweenness_centrality: '0.101515151515'},
+  {source: nodes[Bob], target: nodes[David], friendship: 'enemies', edge_betweenness_centrality: '0.0388888888889'},
+  {source: nodes[Bob], target: nodes[Carlos], friendship: 'enemies', edge_betweenness_centrality: '0.0252525252525'},
+  {source: nodes[David], target: nodes[Carlos], friendship: 'enemies', edge_betweenness_centrality: '0.0489898989899'},
+];
