+++ scripts/new_friends_nodes_nodes.js
@@ -1,0 +1,15 @@
+var SVGTitle = "new friends nodes Hive Panel"
+var nodes = [
+    {name: 'Vivian', gender: ' girl', height: '43', degree_centrality: '0.272727272727', clustering: '0.333333333333', betweenness_centrality: '0.0363636363636', closeness_centrality: '0.458333333333', component_membership: '0'},
+    {name: 'Rob', gender: ' alien', height: '52', degree_centrality: '0.454545454545', clustering: '0.3', betweenness_centrality: '0.274848484848', closeness_centrality: '0.647058823529', component_membership: '0'},
+    {name: 'Alice', gender: ' girl', height: '50', degree_centrality: '0.363636363636', clustering: '0.333333333333', betweenness_centrality: '0.224242424242', closeness_centrality: '0.611111111111', component_membership: '0'},
+    {name: 'George', gender: ' alien', height: '61', degree_centrality: '0.363636363636', clustering: '0.666666666667', betweenness_centrality: '0.0360606060606', closeness_centrality: '0.55', component_membership: '0'},
+    {name: 'Asma', gender: ' girl', height: '47', degree_centrality: '0.0909090909091', clustering: '0.0', betweenness_centrality: '0.0', closeness_centrality: '0.407407407407', component_membership: '0'},
+    {name: 'Jack', gender: ' alien', height: '57', degree_centrality: '0.636363636364', clustering: '0.285714285714', betweenness_centrality: '0.389696969697', closeness_centrality: '0.733333333333', component_membership: '0'},
+    {name: 'Isaac', gender: ' boy', height: '55', degree_centrality: '0.0909090909091', clustering: '0.0', betweenness_centrality: '0.0', closeness_centrality: '0.44', component_membership: '0'},
+    {name: 'Alexandra', gender: ' girl', height: '45', degree_centrality: '0.181818181818', clustering: '1.0', betweenness_centrality: '0.0', closeness_centrality: '0.423076923077', component_membership: '0'},
+    {name: 'Bob', gender: ' boy', height: '53', degree_centrality: '0.454545454545', clustering: '0.6', betweenness_centrality: '0.0587878787879', closeness_centrality: '0.578947368421', component_membership: '0'},
+    {name: 'David', gender: ' boy', height: '49', degree_centrality: '0.272727272727', clustering: '0.666666666667', betweenness_centrality: '0.0121212121212', closeness_centrality: '0.478260869565', component_membership: '0'},
+    {name: 'Eric', gender: ' alien', height: '60', degree_centrality: '0.272727272727', clustering: '0.333333333333', betweenness_centrality: '0.0666666666667', closeness_centrality: '0.52380952381', component_membership: '0'},
+    {name: 'Carlos', gender: ' boy', height: '56', degree_centrality: '0.363636363636', clustering: '0.666666666667', betweenness_centrality: '0.0284848484848', closeness_centrality: '0.52380952381', component_membership: '0'},
+];
